pyproject.toml
[tool.poetry]
name = "openapy"
version = "0.1.0"
description = ""
authors = ["Serj Zavadsky <fevral13@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
PyYAML = "^6.0"
jupyter = "^1.0.0"
mypy = "^0.971"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
